+++ shoop/campaigns/apps.py
@@ -5,7 +5,12 @@
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
+
+from django.db.models.signals import post_save
+
 from shoop.apps import AppConfig
+from shoop.campaigns.signal_handlers import update_customers_groups
+from shoop.core.models import Payment
 
 
 class CampaignAppConfig(AppConfig):
@@ -13,6 +18,9 @@ class CampaignAppConfig(AppConfig):
     verbose_name = "Shoop Campaigns"
     label = "campaigns"
     provides = {
+        "admin_contact_group_form_part": [
+            "shoop.campaigns.admin_module.form_parts:SalesRangesFormPart"
+        ],
         "discount_module": [
             "shoop.campaigns.modules:CatalogCampaignModule"
         ],
@@ -39,3 +47,9 @@ class CampaignAppConfig(AppConfig):
             "shoop.campaigns.models.basket_conditions:ContactBasketCondition",
         ]
     }
+
+    def ready(self):
+        post_save.connect(
+            update_customers_groups,
+            sender=Payment,
+            dispatch_uid="contact_group_sales:update_customers_groups")